remove model and add viewModel
